% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollutionmap.R
\name{APmap}
\alias{APmap}
\title{calculate air pollution map, only used in the project (not going to publish)}
\usage{
APmap(dir = "C:/Users/Lu000012/Documents/files/trajectory/NO2_Input/")
}
\arguments{
\item{dir}{the directory folder of NO2_input}
}
\description{
calculate air pollution map, only used in the project (not going to publish)
}
