% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawtransectline.R
\name{drawtransectline}
\alias{drawtransectline}
\title{plot transect lines on the raster map}
\usage{
drawtransectline(ap, tra_WE = NA, tra_NS = NA)
}
\arguments{
\item{ap}{the raster map where transects are drawn on}

\item{tra_WE}{West-east line, matrix:rbind(c(xmin,y), c(xmax, y))}

\item{tran_ns}{North-South line, matrix}

\item{direc}{which direction to plot, "west-east" or "north-south"}
}
\description{
plot transect lines on the raster map
}
