% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottransec.R
\name{plottransec}
\alias{plottransec}
\title{plot transects}
\usage{
plottransec(ap, tra_WE = NA, tra_NS = NA, direc = c("west-east",
  "north-south"))
}
\arguments{
\item{ap}{the raster map where transects are drawn on}

\item{tra_WE}{West-east line, matrix:rbind(c(xmin,y), c(xmax, y))}

\item{direc}{which direction to plot, "west-east" or "north-south"}

\item{tran_ns}{North-South line, matrix}
}
\description{
plot transects
}
