% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_compare_methods.R
\name{scatter_compare_methods}
\alias{scatter_compare_methods}
\title{Compare Extract location values from raster results and plot the scatterplot}
\usage{
scatter_compare_methods(rasterlist, location, varnames = c("Homemaker",
  "Commuter", "Static"))
}
\arguments{
\item{rasterlist}{list of rasters for comparision}

\item{location}{used \code{\link[raster]{extract}} function, could be e.g. a dataframe with XY to extract locations for comparison.}
}
\description{
Compare Extract location values from raster results and plot the scatterplot
}
