% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_sampling.R
\name{stratified_sampling}
\alias{stratified_sampling}
\title{stratified sampling: divde a raster into grids and ramdomly draw similar amount of data from each grid. The sampled points are moved to the closest target points. In our exposure study, the sampled points are then moved to the nearest home locations.}
\usage{
stratified_sampling(ap, nrofcells = 80, totalpoints = 1020,
  targetpoints, csv = T, result_name = "sample_1000.csv")
}
\arguments{
\item{ap}{rastermap}

\item{nrofcells}{number of grids to be stratified into}

\item{totalpoints}{total points to be drawn, is an approximation since the points are later moved to the closest target points}

\item{targetpoints}{the target points location to move to}

\item{csv}{if T, store with csv file}

\item{result_name}{the filename/directory to store the csv file.}
}
\value{
sampled points
}
\description{
stratified sampling: divde a raster into grids and ramdomly draw similar amount of data from each grid. The sampled points are moved to the closest target points. In our exposure study, the sampled points are then moved to the nearest home locations.
}
